demonstrating synchronization using pyton's queue api

+++ thumbnail_maker.py
@@ -4,7 +4,8 @@ import os
 import logging
 from urllib.parse import urlparse
 from urllib.request import urlretrieve
-import threading
+from queue import Queue
+from threading import Thread
 
 import PIL
 from PIL import Image
@@ -17,55 +18,28 @@ class ThumbnailMakerService(object):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming'
         self.output_dir = self.home_dir + os.path.sep + 'outgoing'
-        self.downloaded_bytes = 0 
-        self.dl_lock = threading.Lock()
-        max_concurrent_dl = 4
-        self.dl_sem = threading.Semaphore(max_concurrent_dl)
-    # working function for the threads
-    def download_image(self, url):
+        self.img_queue = Queue()
+        self.dl_queue = Queue()
 
-        # download each image and save to the input dir 
-        with self.dl_sem:
-            logging.info("downloading image at URL " + url)
-            img_filename = urlparse(url).path.split('/')[-1]
-            dest_path = self.input_dir + os.path.sep + img_filename
-            urlretrieve(url, dest_path)
-            img_size = os.path.getsize(dest_path)
-            # The downloaded_bytes is a sharde recource 
-            # its important to understant that the += operation is three operations that are not atomic 
-            # in the python interpeter as it first reads the value of downloaded_bytes 
-            # then the value of img_size and then stores the sum in downloaded_bytes
-            # so between these operations if there is a contaxt switch then the updaded 
-            # value might be curuppted thus we must use syncrynezation tules such as locks
-            with self.dl_lock:
-                self.downloaded_bytes += img_size
-            logging.info("image [{} bytes] saved to {}".format(img_size, dest_path))
-    
-    def download_images(self, img_url_list):
-        # validate inputs
-        if not img_url_list:
-            return
-        os.makedirs(self.input_dir, exist_ok=True)
-        
-        logging.info("beginning image downloads")
+    def download_image(self):
+        while not self.dl_queue.empty():
+            try:
+                url = self.dl_queue.get(block=False)
+                # download each image and save to the input dir
+                img_filename = urlparse(url).path.split('/')[-1]
+                logging.info("beginning image downloads")
 
-        start = time.perf_counter()
-        threads = []
-        for url in img_url_list:
-            t = threading.Thread(target=self.download_image, args=(url,))
-            t.start()
-            threads.append(t)
-        for t in threads:
-            t.join()
+                urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+                self.img_queue.put(img_filename)
+                self.dl_queue.task_done()
+                logging.info("ending image downloads")
 
-        end = time.perf_counter()
+            except Queue.Empty:
+                logging.info("Queue empty")
 
-        logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
 
     def perform_resizing(self):
         # validate inputs
-        if not os.listdir(self.input_dir):
-            return
         os.makedirs(self.output_dir, exist_ok=True)
 
         logging.info("beginning image resizing")
@@ -73,22 +47,30 @@ class ThumbnailMakerService(object):
         num_images = len(os.listdir(self.input_dir))
 
         start = time.perf_counter()
-        for filename in os.listdir(self.input_dir):
-            orig_img = Image.open(self.input_dir + os.path.sep + filename)
-            for basewidth in target_sizes:
-                img = orig_img
-                # calculate target height of the resized image to maintain the aspect ratio
-                wpercent = (basewidth / float(img.size[0]))
-                hsize = int((float(img.size[1]) * float(wpercent)))
-                # perform resizing
-                img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
-                
-                # save the resized image to the output dir with a modified file name 
-                new_filename = os.path.splitext(filename)[0] + \
-                    '_' + str(basewidth) + os.path.splitext(filename)[1]
-                img.save(self.output_dir + os.path.sep + new_filename)
+        while True:
+            filename = self.img_queue.get()
+            if filename:
+                logging.info("resizing image {}".format(filename))
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
+                    
+                    # save the resized image to the output dir with a modified file name 
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
 
-            os.remove(self.input_dir + os.path.sep + filename)
+                os.remove(self.input_dir + os.path.sep + filename)
+                logging.info("done resizing image {}".format(filename))
+                self.img_queue.task_done()
+            else:
+                self.img_queue.task_done()
+                break
         end = time.perf_counter()
 
         logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
@@ -97,8 +79,24 @@ class ThumbnailMakerService(object):
         logging.info("START make_thumbnails")
         start = time.perf_counter()
 
-        self.download_images(img_url_list)
-        # self.perform_resizing()
+        for img_url in img_url_list:
+            self.dl_queue.put(img_url)
+
+
+        num_dl_threads = 4
+        for _ in range(num_dl_threads):
+            t = Thread(target=self.download_image)
+            t.start()
+
+            
+        t2 = Thread(target=self.perform_resizing)
+        t2.start()
+
+        self.dl_queue.join()
+        self.img_queue.put(None)
+
+        t2.join()
+
 
         end = time.perf_counter()
         logging.info("END make_thumbnails in {} seconds".format(end - start))
